Split Android AI bootstrap provisioning tests
